cleanup, re-use code for loading normalisers, add listnormalisers to api
